+++ src/base_test.go
@@ -4,7 +4,6 @@ package rivescript
 // public facing API from the root rivescript-go package.
 
 import (
-	"fmt"
 	"testing"
 
 	"github.com/aichaos/rivescript-go/sessions"
@@ -31,15 +30,8 @@ func NewTestWithConfig(t *testing.T, debug, utf8 bool, ses sessions.SessionManag
 		t:        t,
 		username: "soandso",
 	}
-	test.bot.Debug = debug
-	test.bot.UTF8 = utf8
-	test.bot.sessions = ses
+	test.bot.Configure(debug, true, utf8, 50, ses)
 	return test
-}
-
-// RS exposes the underlying RiveScript API.
-func (rst *RiveScriptTest) RS() *RiveScript {
-	return rst.bot
 }
 
 // extend updates the RiveScript source code.
@@ -49,17 +41,53 @@ func (rst RiveScriptTest) extend(code string) {
 }
 
 // reply asserts that a given input gets the expected reply.
-func (rst RiveScriptTest) reply(message string, expected string) {
-	reply := rst.bot.Reply(rst.username, message)
-	if reply != expected {
-		rst.t.Error(fmt.Sprintf("Expected %s, got %s", expected, reply))
+func (rst RiveScriptTest) reply(message, expected string) {
+	reply, err := rst.bot.Reply(rst.username, message)
+	if err != nil {
+		rst.t.Errorf("Got an error when checking a reply to '%s': %s", message, err)
+	} else if reply != expected {
+		rst.t.Errorf("Expected %s, got %s", expected, reply)
 	}
 }
 
-// uservar asserts a user variable.
+// replyError asserts that a given input gives an error.
+func (rst RiveScriptTest) replyError(message string, expected error) {
+	if reply, err := rst.bot.Reply(rst.username, message); err == nil {
+		rst.t.Errorf(
+			"Reply to '%s' was expected to error; but it returned %s",
+			message,
+			reply,
+		)
+	} else if err != expected {
+		rst.t.Errorf(
+			"Reply to '%s' got different error than expected; wanted %s, got %s",
+			message,
+			expected,
+			err,
+		)
+	}
+}
+
+// assertEqual checks if two strings are equal.
+func (rst RiveScriptTest) assertEqual(a, b string) {
+	if a != b {
+		rst.t.Errorf("assertEqual: %s != %s", a, b)
+	}
+}
+
+// uservar asserts a user variable is defined and has the expected value.
 func (rst RiveScriptTest) uservar(name string, expected string) {
-	value, _ := rst.bot.GetUservar(rst.username, name)
-	if value != expected {
-		rst.t.Error(fmt.Sprintf("Uservar %s expected %s, got %s", name, expected, value))
+	value, err := rst.bot.GetUservar(rst.username, name)
+	if err != nil {
+		rst.t.Errorf("Got an error when asserting variable %s: %s", name, err)
+	} else if value != expected {
+		rst.t.Errorf("Uservar %s expected %s, got %s", name, expected, value)
+	}
+}
+
+// undefined asserts that a user variable is not set.
+func (rst RiveScriptTest) undefined(name string) {
+	if value, err := rst.bot.GetUservar(rst.username, name); err == nil {
+		rst.t.Errorf("Uservar %s was expected to be undefined; but was %s", name, value)
 	}
 }