+++ src/config.go
@@ -9,6 +9,7 @@ import (
 	"github.com/aichaos/rivescript-go/sessions"
 )
 
+// SetHandler registers a handler for foreign language object macros.
 func (rs *RiveScript) SetHandler(lang string, handler macro.MacroInterface) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
@@ -16,13 +17,23 @@ func (rs *RiveScript) SetHandler(lang string, handler macro.MacroInterface) {
 	rs.handlers[lang] = handler
 }
 
+// RemoveHandler deletes support for a foreign language object macro.
 func (rs *RiveScript) RemoveHandler(lang string) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
 
+	// Purge all loaded objects for this handler.
+	for name, language := range rs.objlangs {
+		if language == lang {
+			delete(rs.objlangs, name)
+		}
+	}
+
+	// And delete the handler itself.
 	delete(rs.handlers, lang)
 }
 
+// SetSubroutine defines a Go function to handle an object macro for RiveScript.
 func (rs *RiveScript) SetSubroutine(name string, fn Subroutine) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
@@ -30,6 +41,7 @@ func (rs *RiveScript) SetSubroutine(name string, fn Subroutine) {
 	rs.subroutines[name] = fn
 }
 
+// DeleteSubroutine deletes a Go object macro handler.
 func (rs *RiveScript) DeleteSubroutine(name string) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
@@ -37,60 +49,67 @@ func (rs *RiveScript) DeleteSubroutine(name string) {
 	delete(rs.subroutines, name)
 }
 
-func (rs *RiveScript) SetGlobal(name string, value string) {
+// SetGlobal configures a global variable in RiveScript.
+func (rs *RiveScript) SetGlobal(name, value string) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
 
-	if value == "undefined" {
+	if value == UNDEFINED {
 		delete(rs.global, name)
 	} else {
 		rs.global[name] = value
 	}
 }
 
-func (rs *RiveScript) SetVariable(name string, value string) {
+// SetVariable configures a bot variable in RiveScript.
+func (rs *RiveScript) SetVariable(name, value string) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
 
-	if value == "undefined" {
-		delete(rs.var_, name)
+	if value == UNDEFINED {
+		delete(rs.vars, name)
 	} else {
-		rs.var_[name] = value
+		rs.vars[name] = value
 	}
 }
 
-func (rs *RiveScript) SetSubstitution(name string, value string) {
+// SetSubstitution sets a substitution pattern.
+func (rs *RiveScript) SetSubstitution(name, value string) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
 
-	if value == "undefined" {
+	if value == UNDEFINED {
 		delete(rs.sub, name)
 	} else {
 		rs.sub[name] = value
 	}
 }
 
-func (rs *RiveScript) SetPerson(name string, value string) {
+// SetPerson sets a person substitution.
+func (rs *RiveScript) SetPerson(name, value string) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
 
-	if value == "undefined" {
+	if value == UNDEFINED {
 		delete(rs.person, name)
 	} else {
 		rs.person[name] = value
 	}
 }
 
-func (rs *RiveScript) SetUservar(username string, name string, value string) {
+// SetUservar sets a user variable to a value.
+func (rs *RiveScript) SetUservar(username, name, value string) {
 	rs.sessions.Set(username, map[string]string{
 		name: value,
 	})
 }
 
+// SetUservars sets many user variables at a time.
 func (rs *RiveScript) SetUservars(username string, data map[string]string) {
 	rs.sessions.Set(username, data)
 }
 
+// GetGlobal retrieves the value of a global variable.
 func (rs *RiveScript) GetGlobal(name string) (string, error) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
@@ -98,51 +117,65 @@ func (rs *RiveScript) GetGlobal(name string) (string, error) {
 	if _, ok := rs.global[name]; ok {
 		return rs.global[name], nil
 	}
-	return "undefined", errors.New("Global variable not found.")
+	return UNDEFINED, errors.New("Global variable not found.")
 }
 
+// GetVariable retrieves the value of a bot variable.
 func (rs *RiveScript) GetVariable(name string) (string, error) {
 	rs.cLock.Lock()
 	defer rs.cLock.Unlock()
 
-	if _, ok := rs.var_[name]; ok {
-		return rs.var_[name], nil
+	if _, ok := rs.vars[name]; ok {
+		return rs.vars[name], nil
 	}
-	return "undefined", errors.New("Variable not found.")
+	return UNDEFINED, errors.New("Variable not found.")
 }
 
-func (rs *RiveScript) GetUservar(username string, name string) (string, error) {
+// GetUservar retrieves the value of a user variable.
+func (rs *RiveScript) GetUservar(username, name string) (string, error) {
 	return rs.sessions.Get(username, name)
 }
 
+// GetUservars retrieves all variables about a user.
 func (rs *RiveScript) GetUservars(username string) (*sessions.UserData, error) {
 	return rs.sessions.GetAny(username)
 }
 
+// GetAllUservars gets all variables about all users.
 func (rs *RiveScript) GetAllUservars() map[string]*sessions.UserData {
 	return rs.sessions.GetAll()
 }
 
+// ClearUservars deletes all the variables that belong to a user.
 func (rs *RiveScript) ClearUservars(username string) {
 	rs.sessions.Clear(username)
 }
 
+// ClearAllUservars deletes all variables for all users.
 func (rs *RiveScript) ClearAllUservars() {
 	rs.sessions.ClearAll()
 }
 
+// FreezeUservars takes a snapshot of a user's variables.
 func (rs *RiveScript) FreezeUservars(username string) error {
 	return rs.sessions.Freeze(username)
 }
 
+// ThawUservars restores a snapshot of user variables.
 func (rs *RiveScript) ThawUservars(username string, action sessions.ThawAction) error {
 	return rs.sessions.Thaw(username, action)
 }
 
+// LastMatch returns the last match of the user.
 func (rs *RiveScript) LastMatch(username string) (string, error) {
 	return rs.sessions.GetLastMatch(username)
 }
 
-func (rs *RiveScript) CurrentUser() string {
-	return rs.currentUser
+// CurrentUser returns the current user and can only be called from within an
+// object macro context.
+func (rs *RiveScript) CurrentUser() (string, error) {
+	if rs.inReplyContext {
+		return rs.currentUser, nil
+	}
+	return "", errors.New("CurrentUser() can only be called inside a reply context")
 }